requirements.txt
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
crispy-bootstrap4==2022.1
cryptography==41.0.1
defusedxml==0.7.1
Django==4.0.3
django-bootstrap4==23.1
django-crispy-forms==2.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-spectacular==0.21.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
oauthlib==3.2.2
Pillow==9.5.0
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.7.0
pyrsistent==0.19.3
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
whitenoise==6.5.0
